backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
distlib==0.3.8
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.3
